requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
django-environ>=0.11,<1.0
django-filter>=24.2,<25.0
whitenoise>=6.6,<7.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
